requirements.txt
python-dotenv
confluent-kafka
pymongo
pytest
pyspark
